+++ src/ui/common/types/ImgComponent.type.ts
